pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "rainier-trader"
version = "0.1.0"
description = "Autonomous AI trading agent for U.S. stocks"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Xue Liu" }]

dependencies = [
    "langgraph>=0.2",
    "anthropic>=0.30",
    "alpaca-py>=0.20",
    "pandas>=2.0",
    "pandas-ta>=0.3",
    "apscheduler>=3.10",
    "typer>=0.12",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]

[project.scripts]
rainier = "rainier_trader.adapters.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["rainier_trader*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
